psi_paramex/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.11.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.2
python-multipart==0.0.6
requests==2.31.0
supabase==2.16.0
pytz==2023.3
jinja2==3.1.2
resend==0.8.0
